+++ src/Zicht/Service/Common/Observers/ServiceObserverAdapter.php
@@ -6,14 +6,14 @@
 namespace Zicht\Service\Common\Observers;
 
 use Zicht\Service\Common\ServiceCallInterface;
-use Zicht\Service\Common\ServiceObserver;
+use Zicht\Service\Common\ServiceObserverInterface;
 
 /**
  * Adapter for the ServiceObserver interface
  *
  * @codeCoverageIgnore
  */
-class ServiceObserverAdapter implements ServiceObserver
+class ServiceObserverAdapter implements ServiceObserverInterface
 {
     /**
      * {@inheritdoc}